fix git ignore

+++ spark-word-count/src/main/scala/WordAggregation.scala
@@ -10,7 +10,7 @@ object WordAggregation {
     val sc= new SparkContext(sparkConf)
 
     val rddList= sc.parallelize(Array(("saprk",3),("hadoop",5),("yarn",3),("saprk",1),("hadoop",2),("yarn",4)))
-     val aggregatedWords= rddList.aggregateByKey(0)((k,v) => v.toInt+k, (v,k) => k+v).collect
+     val aggregatedWords= rddList.aggregateByKey(0)((k,v) => v+k, (v,k) => k+v).collect
     println("result of aggregate By Key" + aggregatedWords.toList)
 
   }
